# api.yaml
openapi: 3.0.3
...
info:
  title: MangaDex API
  version: 5.10.2
  contact:
    name: MangaDex staff team
    email: "support@mangadex.org"
  description: |
    MangaDex is an ad-free manga reader offering high-quality images!

    This document details our API as it is right now. It is in no way a promise to never change it, although we will endeavour to publicly notify any major change.

    # Acceptable use policy

    Usage of our services implies acceptance of the following:
    - You **MUST** credit us
    - You **MUST** credit scanlation groups if you offer the ability to read chapters
    - You **CANNOT** run ads or paid services on your website and/or apps

    These may change at any time for any and no reason and it is up to you check for updates from time to time.

    # Security issues

    If you believe you found a security issue in our API, please check our [security.txt](/security.txt) to get in touch privately.
servers:
  - url: 'https://api.mangadex.org'
    description: Production
  - url: 'https://api.mangadex.dev'
    description: Public development environment
paths:
...
  '/manga/{id}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Manga ID
    get:
      summary: Get Manga
      tags:
        - Manga
      parameters:
        - schema:
            $ref: '#/components/schemas/ReferenceExpansionManga'
          in: query
          name: includes[]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Manga no content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get-manga-id
      description: Get Manga.
      security: []
    put:
      summary: Update Manga
      operationId: put-manga-id
      parameters:
        - schema:
            type: string
            default: application/json
          name: Content-Type
          in: header
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MangaResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/MangaEdit'
                - type: object
                  properties:
                    artists:
                      type: array
                      items:
                        type: string
                        format: uuid
                        minLength: 36
                        maxLength: 36
                    authors:
                      type: array
                      items:
                        type: string
                        format: uuid
                        minLength: 36
                        maxLength: 36
        description: The size of the body is limited to 64KB.
      tags:
        - Manga
    delete:
      summary: Delete Manga
      operationId: delete-manga-id
      responses:
        '200':
          description: Manga has been deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Manga
...
  /settings/template:
    get:
      summary: Get latest Settings template
      operationId: get-settings-template
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                description: JSON Schema to validate settings
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Manga no content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Settings
    post:
      summary: Create Settings template
      operationId: post-settings-template
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: A JSON Schema to validate settings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                description: JSON Schema to validate settings
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Settings
  /settings/template/{version}:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
          name: version
          in: path
          required: true
      summary: Get Settings template by version id
      operationId: get-settings-template-version
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                description: JSON Schema to validate settings
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Manga no content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Settings
...
components:
  schemas:
    MangaRequest:
      description: ''
      type: object
      title: MangaRequest
      properties:
        title:
          $ref: '#/components/schemas/LocalizedString'
        altTitles:
          type: array
          items:
            $ref: '#/components/schemas/LocalizedString'
        description:
          $ref: '#/components/schemas/LocalizedString'
        authors:
          type: array
          items:
            type: string
            format: uuid
        artists:
          type: array
          items:
            type: string
            format: uuid
        links:
          type: object
          additionalProperties:
            type: string
        originalLanguage:
          type: string
          pattern: '^[a-z]{2}(-[a-z]{2})?$'
        lastVolume:
          type: string
          nullable: true
        lastChapter:
          type: string
          nullable: true
        publicationDemographic:
          type: string
          nullable: true
          enum:
            - shounen
            - shoujo
            - josei
            - seinen
        status:
          type: string
          nullable: false
          enum:
            - completed
            - ongoing
            - cancelled
            - hiatus
        year:
          type: integer
          nullable: true
          maximum: 9999
          minimum: 1
          description: Year of release
        contentRating:
          type: string
          enum:
            - safe
            - suggestive
            - erotica
            - pornographic
        chapterNumbersResetOnNewVolume:
          type: boolean
        tags:
          type: array
          items:
            type: string
            format: uuid
        primaryCover:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          minimum: 1
    LocalizedString:
      type: object
      title: LocalizedString
      additionalProperties:
        type: string
        pattern: "^[a-z]{2,8}$"
    MangaResponse:
      title: MangaResponse
      type: object
      properties:
        result:
          type: string
          enum:
            - ok
            - error
        response:
          type: string
          default: entity
        data:
          $ref: '#/components/schemas/Manga'
...
    Relationship:
      title: Relationship
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        related:
          type: string
          enum: ['monochrome', 'main_story', 'adapted_from', 'based_on', 'prequel', 'side_story', 'doujinshi', 'same_franchise', 'shared_universe', 'sequel', 'spin_off', 'alternate_story', 'alternate_version', 'preserialization', 'colored', 'serialization']
          description: Related Manga type, only present if you are on a Manga entity and a Manga relationship
        attributes:
          description: If Reference Expansion is applied, contains objects attributes
          nullable: true
          type: object
...
    Manga:
      title: Manga
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - manga
        attributes:
          $ref: '#/components/schemas/MangaAttributes'
        relationships:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
    ErrorResponse:
      title: ErrorResponse
      type: object
      properties:
        result:
          type: string
          default: error
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    Error:
      title: Error
      type: object
      properties:
        id:
          type: string
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
          nullable: true
        context:
          type: string
          nullable: true
...
    MangaAttributes:
      title: MangaAttributes
      type: object
      properties:
        title:
          $ref: '#/components/schemas/LocalizedString'
        altTitles:
          type: array
          items:
            $ref: '#/components/schemas/LocalizedString'
        description:
          $ref: '#/components/schemas/LocalizedString'
        isLocked:
          type: boolean
        links:
          type: object
          additionalProperties:
            type: string
        originalLanguage:
          type: string
        lastVolume:
          type: string
          nullable: true
        lastChapter:
          type: string
          nullable: true
        publicationDemographic:
          nullable: true
          type: string
          enum:
            - shounen
            - shoujo
            - josei
            - seinen
        status:
          type: string
          nullable: false
          enum:
            - completed
            - ongoing
            - cancelled
            - hiatus
        year:
          type: integer
          nullable: true
          description: Year of release
        contentRating:
          type: string
          enum:
            - safe
            - suggestive
            - erotica
            - pornographic
        chapterNumbersResetOnNewVolume:
          type: boolean
        availableTranslatedLanguages:
          type: array
          items:
            type: string
            pattern: '^[a-z]{2}(-[a-z]{2})?$'
        latestUploadedChapter:
          type: string
          format: uuid
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        state:
          type: string
          enum: ['draft', 'submitted', 'published', 'rejected']
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
        updatedAt:
          type: string
...
    MangaEdit:
      allOf:
        - $ref: '#/components/schemas/MangaRequest'
        - required:
            - version
...
    Response:
      title: Response
      type: object
      properties:
        result:
          type: string
          enum:
            - ok
            - error
...
    Tag:
      title: Tag
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - tag
        attributes:
          $ref: '#/components/schemas/TagAttributes'
        relationships:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
    TagAttributes:
      title: TagAttributes
      type: object
      properties:
        name:
          $ref: '#/components/schemas/LocalizedString'
        description:
          $ref: '#/components/schemas/LocalizedString'
        group:
          type: string
          enum:
            - content
            - format
            - genre
            - theme
        version:
          type: integer
          minimum: 1
...
    ReferenceExpansionManga:
      title: ReferenceExpansionManga
      description: Reference expansion options for manga entities or lists
      type: array
      items:
        type: string
        enum:
          - manga
          - cover_art
          - author
          - artist
          - tag
          - creator
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
security:
  - Bearer: []